Updated test project structure
