+++ Tests/CMakeLists.txt
@@ -7,11 +7,16 @@ include(Doctest)
 
 set(SOURCE_LIST
     ${PROJECT_SOURCE_DIR}/Main.cpp
-    ${PROJECT_SOURCE_DIR}/RTProtocolXmlTests.cpp
-    ${PROJECT_SOURCE_DIR}/RTProtocolSkeletonSettingsXmlTests.cpp
-    ${PROJECT_SOURCE_DIR}/GazeVectorParameterXmlTests.cpp
     ${PROJECT_SOURCE_DIR}/CalibrationXmlTests.cpp
-
+    ${PROJECT_SOURCE_DIR}/EyeTrackerXmlTests.cpp
+    ${PROJECT_SOURCE_DIR}/GazeVectorXmlTests.cpp
+    ${PROJECT_SOURCE_DIR}/GeneralXmlTests.cpp
+    ${PROJECT_SOURCE_DIR}/ImageXmlTests.cpp
+    ${PROJECT_SOURCE_DIR}/SkeletonXmlTests.cpp
+    ${PROJECT_SOURCE_DIR}/3dXmlTests.cpp
+    ${PROJECT_SOURCE_DIR}/6dXmlTests.cpp
+    ${PROJECT_SOURCE_DIR}/AnalogXmlTests.cpp
+    ${PROJECT_SOURCE_DIR}/ForceXmlTests.cpp
 )
 
 add_executable(
