+++ Tests/GazeVectorXmlTests.cpp
@@ -1,10 +1,5 @@
-#include "TestUtils.h"
-#include "../RTProtocol.h"
-#include "Data/EyeTracker.h"
-
-#include "Data/Gaze.h"
-#include "doctest/doctest.h"
-#include <memory>
+#include "Data/GazeVector.h"
+#include "XmlTestsShared.h"
 
 using namespace qualisys_cpp_sdk::tests;
 
@@ -32,26 +27,4 @@ TEST_CASE("GetSettingsGazeVectorTest")
     CHECK_EQ(240.0f, gazeVectorSettings[1].frequency);
     CHECK_EQ(true, gazeVectorSettings[1].hwSync);
     CHECK_EQ(false, gazeVectorSettings[1].filter);
-}
-
-TEST_CASE("GetSettingsEyeTrackerTest")
-{
-    auto [protocol, network] = utils::CreateTestContext();
-
-    network->PrepareResponse("GetParameters EyeTracker", data::GetEyeTrackerSettingsTest, CRTPacket::PacketXML);
-
-    bool bDataAvailable = true;
-
-    if (!protocol->ReadEyeTrackerSettings(bDataAvailable))
-    {
-        FAIL(protocol->GetErrorString());
-    }
-
-    std::vector<CRTProtocol::SEyeTracker> eyeTrackerSettings;
-    protocol->GetEyeTrackerSettings(eyeTrackerSettings);
-
-    CHECK_EQ(1, eyeTrackerSettings.size());
-    CHECK_EQ("EyeTrackerDevice", eyeTrackerSettings[0].name);
-    CHECK_EQ(240.0f, eyeTrackerSettings[0].frequency);
-    CHECK_EQ(true, eyeTrackerSettings[0].hwSync);
 }
