+++ Tests/XmlTestsShared.h
@@ -1,0 +1,9 @@
+#pragma once
+
+#include "TestUtils.h"
+
+#include <doctest/doctest.h>
+
+#include <Network.h>
+#include <RTProtocol.h>
+#include <RTPacket.h>
